revert the row_num stuff that was put in for this issue. #176
Change the pop method so that it pops off of the back of the list
instead of the front.

The row_num can be added to the Track bay calling len(playlist). This is
implemented in __load_last_row_num(track).

The row nm entries are removed from the teo column modules.