create and implement book_ease_tables._LoaderInterface closes #300
Create 2 implementations of the interface. _LoaderSettingsNumeric and
_LoaderSettingsString.

add 2 functions to book_ease_tables that orchestrate the loading,
copying, and saving of the book_ease.db

load_data()
Calls on the various _LoaderInteraces to copy the on disk database to
the in mem database.

save_data()
Calls on the various _LoaderInteraces to copy the in mem database to the
on disk database.

add _init_database()
initializes the tables of a database

_attach_book_ease_db():
Attach the on disk database to the connection to the in-memory database

_detach_book_ease_db():
Detach the on disk database from the connection to the in-memory
database

call book_ease_tables.save_data() and load_data() from book_ease.main.