make the has_playlist combo box resizable close #556
file_mgr.ui:
Move the open button so that it is next to the create button instead of
on the other side of the existing playlist combobox.

file_mgr.py:
set the width-chars (minimum width) porperty on the Gtk.CellRendererText
that underlies the combo box.
This also requires setting the ellipsize property to
gi.repository.Pango.EllipsizeMode.END